fix linter issues and fix proptypes issues in BooksList
