+++ src/reducers/books.js
@@ -1,6 +1,7 @@
 import { CREATE_BOOK, REMOVE_BOOK } from '../actions/index';
+import defaultBooks from './index';
 
-const booksReducer = (state, action) => {
+const booksReducer = (state = defaultBooks, action) => {
   switch (action.type) {
     case CREATE_BOOK:
       return [...state, action.payload];
