+++ src/reducers/index.js
@@ -1,0 +1,22 @@
+const defaultBooks = [
+  {
+    id: Math.floor(Math.random() * 1000),
+    title: 'The war through the history',
+    category: 'History',
+
+  },
+  {
+    id: Math.floor(Math.random() * 1000),
+    title: 'Harry Potter',
+    category: 'Action',
+
+  },
+  {
+    id: Math.floor(Math.random() * 1000),
+    title: 'life of Bill Gates',
+    category: 'Biography',
+
+  },
+];
+
+export default defaultBooks;
